fix category

+++ src/services/action.js
@@ -1,6 +1,7 @@
 const CustomError = require('../errors/CustomError');
 const errorCodes = require('../errors/code');
 const actionDao = require('../daos/action');
+const nodeDao = require('../daos/node');
 
 const findAllActionByBotId = async ({ botId, fields, sort }) => {
   // const newFields = fields.split(',');
@@ -69,6 +70,10 @@ const updateAction = async ({ id, name, actions, groupActionId, botId }) => {
 };
 
 const deleteAction = async (id) => {
+  const node = await nodeDao.findNodeByCondition({ action: id });
+  if (node) {
+    throw CustomError(errorCodes.ITEM_EXIST_IN_WORKFLOW);
+  }
   await actionDao.deleteAction(id);
 };
